Add support for compiling GFXStream
This add support for compiling the GFXStream dependency using cmake,
this is a requirement for getting rutabaga to work.

+++ google/compat/windows/includes/sys/types.h
@@ -10,5 +10,6 @@
 // GNU General Public License for more details.
 #pragma once
 #include_next <sys/types.h>
+#include <inttypes.h>
 
-typedef int pid_t;
+typedef int64_t pid_t;
